Move ComponentRecord to component manager
