+++ src/index.ts
@@ -20,7 +20,8 @@ export { createWorld } from "./world.js";
 
 export type { Archetype } from "./archetype/archetype.js";
 export type { Bitfield } from "./bitfield.js";
-export type { Component, ComponentRecord, ComponentSpec } from "./component/component.js";
+export type { Component, ComponentSpec } from "./component/component.js";
+export type { ComponentRecord } from "./component/manager.js";
 export type { ComponentInstance } from "./component/instance.js";
 export type { StorageProxy } from "./component/proxy.js";
 export type { Entity } from "./entity.js";
